% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_calcs.R
\name{coh}
\alias{coh}
\title{Cohen's d}
\usage{
coh(n1, n2, mn1, mn2, vr1, vr2)
}
\arguments{
\item{n1}{The sample size for group 1}

\item{n2}{The sample size for group 2}

\item{mn1}{The mean for group 1}

\item{mn2}{The mean for group 2}

\item{vr1}{The variance for group 1}

\item{vr2}{The variance for group 2}
}
\description{
Wraps the equation into a function
}
\keyword{internal}
