% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{col_scheme}
\alias{col_scheme}
\title{Determine the color scheme to be used for the plotting}
\usage{
col_scheme(scheme, n, ...)
}
\arguments{
\item{scheme}{The chosen color scheme. Options are "ggplot2", "viridis",
"magma", "inferno", or "plasma". Note all but "ggplot2" depend upon the
\href{https://CRAN.R-project.org/package=viridisLite}{viridisLite} package}

\item{n}{The number of colors to be produced for the given scheme.}

\item{...}{Additional arguments passed, typically being \code{alpha}.}
}
\description{
Determine the color scheme to be used for the plotting
}
