% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{create_base_legend}
\alias{create_base_legend}
\title{Create a base legend for a plot}
\usage{
create_base_legend(labels, position = "bottomright", ...)
}
\arguments{
\item{labels}{Labels for the legend (line labels).}

\item{position}{Where the legend should be positioned. Defaults to
\code{"bottomright"}.}

\item{...}{Additional arguments passed to \link[graphics]{legend} 
(typically colors, line width, etc).}
}
\description{
This function creates a legend using the base \link[graphics]{legend}
function, but with more abbreviated syntax.
}
