% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_utils.R
\name{create_cut_refs}
\alias{create_cut_refs}
\title{Create a set of reference lines according to a cut score}
\usage{
create_cut_refs(cut, calcs, p, scheme)
}
\arguments{
\item{cut}{The cut scores on the raw scale}

\item{calcs}{object from \link{pp_calcs}}

\item{p}{output from \link{empty_plot}}

\item{scheme}{What color scheme should the lines follow? Defaults to 
mimic the ggplot2 color scheme. Other options come from the 
\href{https://CRAN.R-project.org/package=viridisLite}{viridisLite}
package, and must be installed first. These are the same options available
in the package: "viridis", "magma", "inferno", and "plasma". These color 
schemes work well for color blindness and print well in black and white.
Alternatively, colors can be supplied manually through a call to \code{col}
(through \code{...}).}
}
\description{
Create a set of reference lines according to a cut score
}
