% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{create_mat}
\alias{create_mat}
\title{Create Matrix of all possible combinations}
\usage{
create_mat(levs, fun, diagonal = 0, vec = FALSE)
}
\arguments{
\item{levs}{The levels of the grouping factor from which to create the
matrix}

\item{fun}{The function to apply.}

\item{diagonal}{What elements should go on the diagonal of the matrix?
Defaults to 0.}

\item{vec}{Logical. Should a vector rather than a matrix be returned? 
Defaults to FALSE.}
}
\value{
Matrix of values according to the function supplied.
}
\description{
Archived Code: Data are now returned in a tidy data format
}

