% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{create_vec}
\alias{create_vec}
\title{Create a named vector of all possible combinations}
\usage{
create_vec(levs, fun)
}
\arguments{
\item{levs}{The levels of the grouping factor from which to create the
matrix}

\item{fun}{The function to apply.}
}
\value{
Matrix of values according to the function supplied.
}
\description{
Alternative to tidied data frame return.
}
