% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{descrip_stats}
\alias{descrip_stats}
\title{Report descriptive stats for all possible pairings on the rhs of the formula.}
\usage{
descrip_stats(data, formula, ..., qtile_groups = NULL)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups. Additional variables 
can be included with \code{+} to produce descriptive stats by the secondary 
or tertiary variable of interest (e.g., \code{out ~ group + characteristic1 
+ characteristic2}).}
}
\description{
Report descriptive stats for all possible pairings on the rhs of the formula.
}
\keyword{internal}
