% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{empty_plot}
\alias{empty_plot}
\title{Create an empty plot}
\usage{
empty_plot(x, y, default_xlab, default_ylab, default_main, default_bty = "n",
  default_xlim = NULL, default_ylim = NULL, ...)
}
\arguments{
\item{x}{The x variable to be plotted.}

\item{y}{The y variable to be plotted.}

\item{default_xlab}{The default x-label, which can be overridden by the 
user.}

\item{default_ylab}{The default y-label, which can be overridden by the 
user.}

\item{default_main}{The default main title, which can be overridden by the 
user.}

\item{default_bty}{The default background type, which can be overridden by 
the user. Defaults to \code{"n"}.}

\item{...}{Additional arguments supplied to \link[graphics]{plot} (e.g., 
\code{xlim}, \code{ylim}, \code{cex}, etc.)}
}
\description{
This function creates an empty plot for further plotting (e.g., via 
\link[graphics]{lines}). What makes the function unique is that it allows
for specification of default \code{xlab}, \code{ylab}, and \code{main}
arguments, while allowing the user to override those arguments. Only really
useful when used within other functions (e.g., \link{pp_plot}).
}

