% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_utils.R
\name{extract_pp_data}
\alias{extract_pp_data}
\title{Produce calculations necessary for \link{pp_plot}.}
\usage{
extract_pp_data(data, formula, ref_group = NULL)
}
\arguments{
\item{data}{The data frame to be plotted}

\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups. Additional variables 
can be included with \code{+} to produce separate plots by the secondary or
tertiary variable of interest (e.g., \code{out ~ group + characteristic1 + 
characteristic2}). No more than two additional characteristics can be 
supplied at this time.}

\item{ref_group}{Optional character vector (of length 1) naming the
reference group. Defaults to the group with the highest mean score.}
}
\value{
A dataframe either for use with plotting or for effect size 
  calculation.
}
\description{
Produce calculations necessary for \link{pp_plot}.
}
\examples{

extract_pp_data(star, math ~ condition)
 
extract_pp_data(benchmarks, math ~ ell + season + frl)

}
