% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_calcs.R
\name{hedg}
\alias{hedg}
\title{Hedge's g}
\usage{
hedg(n1, n2, d)
}
\arguments{
\item{n1}{The sample size for group 1}

\item{n2}{The sample size for group 2}

\item{d}{The value of Cohen's d}
}
\description{
Wraps the equation into a function
}
\keyword{internal}
