% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{hedg_g}
\alias{hedg_g}
\title{Compute Hedges' \emph{g}
This function calculates effect sizes in terms of Hedges' \emph{g}, also
called the corrected (for sample size) effect size. See
\code{\link{coh_d}} for the uncorrected version. Also see 
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3840331/}{Lakens (2013)}
for a discussion on different types of effect sizes and their
interpretation. Note that missing data are removed from the calculations of 
the means and standard deviations.}
\usage{
hedg_g(formula, data, matrix = TRUE)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{matrix}{Logical. If only two groups are being compared, should the
full matrix of all possible comparisons be returned? If \code{FALSE} only
the positive effect size is returned.}
}
\value{
By default the Hedges' \emph{d} for all possible pairings of
 the grouping factor are returned as a matrix, with the reference group 
 reported by rows and the focal group reported by columns.
}
\description{
Compute Hedges' \emph{g}
This function calculates effect sizes in terms of Hedges' \emph{g}, also
called the corrected (for sample size) effect size. See
\code{\link{coh_d}} for the uncorrected version. Also see 
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3840331/}{Lakens (2013)}
for a discussion on different types of effect sizes and their
interpretation. Note that missing data are removed from the calculations of 
the means and standard deviations.
}

