% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{parse_form}
\alias{parse_form}
\title{Parse formula}
\usage{
parse_form(formula, data, order = TRUE)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{order}{Logical. Defaults to \code{TRUE}. Should the groups be ordered 
according to their mean?}
}
\value{
A list of data split by the grouping factor.
}
\description{
Parse formula
}
