% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_utils.R
\name{pooled_sd}
\alias{pooled_sd}
\title{Compute pooled standard deviation}
\usage{
pooled_sd(formula, data)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note the
grouping variable must only include only two groups.}

\item{data}{The data frame that the data in the formula come from.}
}
\description{
Compute pooled standard deviation
}
\examples{
pooled_sd(math ~ condition, star)
pooled_sd(reading ~ sex, star)
}
