% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_utils.R
\name{pp_annotate}
\alias{pp_annotate}
\title{Annotation function to add AUC/V to a given plot}
\usage{
pp_annotate(formula, data, ref_group = NULL, x, y, text_size,
  theme = "standard")
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note the
grouping variable must only include only two groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{ref_group}{Optional character vector (of length 1) naming the
reference group to be plotted on the x-axis. Defaults to the highest
scoring group.}

\item{x}{The x-axis location for the text.}

\item{y}{The y-axis location for the text.}

\item{text_size}{Size of the text used in the annotation}

\item{theme}{The theme for the plot.}
}
\description{
Annotation function to add AUC/V to a given plot
}
