% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_plot.R
\name{pp_plot}
\alias{pp_plot}
\title{Produces the paired probability plot for two groups}
\usage{
pp_plot(formula, data, ref_group = NULL, refline = TRUE,
  refline_col = "gray", refline_lty = 2, refline_lwd = 1, text = NULL,
  shade = NULL, shade_rgb = rgb(102, 178, 255, alpha = 30, maxColorValue =
  255), legend = NULL, return = FALSE, plot = TRUE, par_reset = TRUE,
  ...)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note the
grouping variable must only include only two groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{ref_group}{Optional character vector (of length 1) naming the
reference group to be plotted on the x-axis. Defaults to the highest
scoring group.}

\item{refline}{Logical. Defaults to \code{TRUE}. Should a diagonal
reference line, representing the point of equal probabilities, be
   plotted?}

\item{refline_col}{Color of the reference line.}

\item{refline_lty}{Line type of the reference line.}

\item{refline_lwd}{Line width of the reference line.}

\item{text}{Logical. Should the \code{link{auc}} and \code{link{v}}
statistics be displayed on the plot? Defaults to \code{TRUE} when there
   are two groups. Cannot currently be displayed for more than two groups.}

\item{shade}{Logical. Should the area under the curve be shaded? Defaults
to \code{TRUE} if there are only two group. Currently it cannot be 
produced for more than two groups.}

\item{shade_rgb}{The color of the shading via \link[grDevices]{rgb}. 
Defaults to \code{rgb(102, 178, 255, alpha = 30, max = 255)} which is a 
light blue color.}

\item{legend}{Logical. Should the legend be plotted? Defaults to \code{TRUE}
when the number of groups is greater than 2. Legend not available when
the number of groups == 2.}

\item{return}{Logical. Should the arguments passed to \link[graphics]{plot}
returned (in quoted form)? Defaults to \code{FALSE}.}

\item{plot}{Logical. Should the plot be produced? Defaults to \code{TRUE}. 
Sometimes it is useful to only get the output from the plot, which is why
this functionality exists (likely to be implemented in a panel plot).}

\item{par_reset}{Logical. Should the \link[graphics]{par} parameters be 
reset after producing the plot? Defaults to \code{TRUE}. Set to 
\code{FALSE} when using the function in multi-panel plots (e.g., when
\code{mfrow} is used.}

\item{...}{Additional arguments passed to \link[graphics]{plot}. Note that
it is best to use the full argument rather than partial matching, given the
method used to call the plot. While some partial matching is supported 
(e.g., \code{m} for \code{main}, it is generally safest to supply the full
argument).}
}
\description{
The paired probability plot maps the probability of obtaining a specific
   score for each of two groups. The area under the curve 
   (\code{\link{auc}}) corresponds to the probability that the x-axis group
	   will score higher than the y-axis group.
}
\examples{
free_reduced <- rnorm(800, 80, 20)
 pay <- rnorm(500, 100, 10)
 d <- data.frame(score = c(free_reduced, pay), 
 				frl = c(rep("free_reduced", 800),  
							rep("pay", 500)))

 pp_plot(score ~ frl, d)
}

