% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_calcs.R
\name{psd}
\alias{psd}
\title{Pooled Standard Deviation}
\usage{
psd(n1, n2, vr1, vr2)
}
\arguments{
\item{n1}{The sample size for group 1}

\item{n2}{The sample size for group 2}

\item{vr1}{The variance for group 1}

\item{vr2}{The variance for group 2}
}
\description{
The denominator for Cohen's d
}
\keyword{internal}
