% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{themes}
\alias{themes}
\title{Theme settings}
\usage{
themes(theme)
}
\arguments{
\item{theme}{The name of the theme.}
}
\value{
list with the \code{par} settings and the primary line color (e.g., 
white for theme = "dark" and black for theme = "standard").
}
\description{
Parameters for each theme currently implemented.
}
