% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{tpac}
\alias{tpac}
\title{Transformed proportion above the cut}
\usage{
tpac(formula, data, cut, groups = "all", diff = TRUE, matrix = TRUE)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{cut}{The point at the scale from which the proportion above should
be calculated from.}

\item{groups}{The groups to return the proportion above the cut. Defaults
to \code{"all"}. Argument takes a character or character vector as its
argument corresponding to the levels of the grouping factor.}

\item{diff}{Logical, defaults to \code{TRUE}. Should the difference between
the groups be returned? If \code{FALSE} the raw transformed proportion above
the cut is returned for each group, in standard deviation units.}

\item{matrix}{Logical, defaults to \code{TRUE}. Should the results be
returned as a matrix? Only relevant when \code{diff == TRUE}.}
}
\value{
Matrix (or vector) of the transformed proportion above the cutoff.
}
\description{
This function transforms calls to \link{pac} into standard deviation units.
Function assumes that each distribution is distributed normally with 
common variances. See 
\href{http://journals.sagepub.com/doi/abs/10.3102/1076998611411918}{Ho &
 Reardon, 2012}
}

