% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{v}
\alias{v}
\title{Calculate the V effect size statistic}
\usage{
v(ps)
}
\arguments{
\item{ps}{The paired probabilities for two groups, calculated from 
\code{\link{probs}}. Note that the probabilities must only be from two
groups.}
}
\description{
This function calculates the effect size V, as discussed by Ho.
}
\examples{
free_reduced <- rnorm(800, 80, 20)
pay <- rnorm(500, 100, 10)
d <- data.frame(score = c(free_reduced, pay), 
				frl = c(rep("free_reduced", 800),  
						rep("pay", 500)))

v(probs(score ~ frl, d))
}

