% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{auc}
\alias{auc}
\title{Calculate the area under the curve}
\usage{
auc(ps)
}
\arguments{
\item{ps}{The paired probabilities for two groups, calculated from 
\code{\link{probs}}. Note that the probabilities must only be from two
groups.}
}
\description{
This function is used within \code{\link{pp_plot}} to calculate the area 
under the \code{pp} curve.
}
\examples{
free_reduced <- rnorm(800, 80, 20)
pay <- rnorm(500, 100, 10)
d <- data.frame(score = c(free_reduced, pay), 
				frl = c(rep("free_reduced", 800),  
						rep("pay", 500)))

auc(probs(score ~ frl, d))
}

