% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_calcs.R
\name{auc}
\alias{auc}
\title{Calculate the area under the curve}
\usage{
auc(formula, data, matrix = TRUE)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{matrix}{Logical. If only two groups are being compared, should the
full matrix of all possible comparisons be returned? If \code{FALSE} the
greater of the two aucs is returned.}
}
\value{
By default the area under the curve for all possible pairings of
 the grouping factor are returned as a matrix, with the reference group 
 (x-axis of the pp plot) reported by rows and the focal group 
 (y-axis) reported by columns.
}
\description{
This function is used within \code{\link{pp_plot}} to calculate the area 
under the \code{pp} curve. The area under the curve is also a useful 
effect-size like statistic, representing the probability that a randomly 
selected individual from distribution a will have a higher value than a 
randomly selected individual from distribution b.
}
\examples{
free_reduced <- rnorm(800, 80, 20)
pay <- rnorm(500, 100, 10)
d <- data.frame(score = c(free_reduced, pay), 
				frl = c(rep("free_reduced", 800),  
						rep("pay", 500)))

auc(score ~ frl, d)
}

