% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{col_hue}
\alias{col_hue}
\title{Color hues}
\usage{
col_hue(n)
}
\arguments{
\item{n}{The number of colors to be produced}
}
\description{
Emulates ggplot's default colors. Evenly spaced hues around the color wheel.
}
\examples{
col_hue(1)
col_hue(5)
col_hue(20)
}
