% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf_plot.R
\name{ecdf_plot}
\alias{ecdf_plot}
\title{Empirical Cumulative Distribution Plot}
\usage{
ecdf_plot(formula, data, v_ref = NULL, ref_groups = NULL, ref_col = NULL,
  text = TRUE, legend = TRUE, return = FALSE, ...)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{v_ref}{Optional argument stating the location of the vertical 
reference line(s) to compare groups at a specific location on the scale.}

\item{ref_groups}{Optional character vector indicating the reference groups
(CDFs) to be compared at the location of \emph{v_ref}.}

\item{ref_col}{Optional argument stating the color of the vertical reference
line. Defaults to gray70.}

\item{text}{Logical. If reference groups are compared, should the difference
between the curves be annotated to the plot along the y-axis?}

\item{legend}{Logical. Should the legend be plotted? Defaults to 
\code{TRUE}.}

\item{return}{Logical. Should the arguments passed to \link[graphics]{plot}
returned (in quoted form)? Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to \link[graphics]{plot}. Note that
it is best to use the full argument rather than partial matching, given the
method used to call the plot. While some partial matching is supported 
(e.g., \code{m} for \code{main}, it is generally safest to supply the full
argument).}
}
\description{
This function dresses up the \link[stats]{plot.ecdf} function and provides
some additional functionality to directly compare distributions at specific
locations along the scale. Specifically, multiple empirical CDFs can be 
plotted with a single call, and the differences between any pair, or all, 
CDFs can optionally be plotted in terms of both raw percentage differences 
and/or in terms of standard deviation units through inverse normal
transformations. See 
\href{http://journals.sagepub.com/doi/abs/10.3102/1076998611411918}{Ho &
 Reardon, 2012}. (Note, not all features implemented yet)
}

