% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_utils.R
\name{sanitize_names}
\alias{sanitize_names}
\title{Internal function to clean/sanitize a character vector of dahses, spaces, 
and periods.}
\usage{
sanitize_names(x)
}
\arguments{
\item{x}{Character vector to be sanitized.}
}
\description{
Internal function to clean/sanitize a character vector of dahses, spaces, 
and periods.
}
\keyword{internal}
