% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_calcs.R, R/helper_funs.R
\name{v}
\alias{v}
\title{Calculate the V effect size statistic}
\usage{
v(formula, data)

v(formula, data)
}
\arguments{
\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups.}

\item{data}{The data frame that the data in the formula come from.}

\item{formula}{A formula of the type \code{out ~ group} where \code{out} is
the outcome variable and \code{group} is the grouping variable. Note this
variable can include any arbitrary number of groups.}

\item{data}{The data frame that the data in the formula come from.}
}
\value{
By default the V statistic for all possible pairings of
 the grouping factor are returned as a matrix, with the reference group 
 (x-axis of the pp plot) reported by rows and the focal group 
 (y-axis) reported by columns. Note that V cannot be negative (given that
 the square root of a negative number is imaginary) and half the values
 reported are missing (typically the lower triangle of the matrix). When
 only two groups are included in the grouping factor, a single value is
 returned.

By default the V statistic for all possible pairings of
 the grouping factor are returned as a matrix, with the reference group 
 (x-axis of the pp plot) reported by rows and the focal group 
 (y-axis) reported by columns. Note that V cannot be negative (given that
 the square root of a negative number is imaginary) and half the values
 reported are missing (typically the lower triangle of the matrix). When
 only two groups are included in the grouping factor, a single value is
 returned.
}
\description{
This function calculates the effect size V, as discussed by Ho, 2009. The V
statistic is a transformation of \code{\link{auc}}, interpreted as the 
average difference between the distributions in standard deviation units.

This function calculates the effect size V, as discussed by Ho, 2009. The V
statistic is a transformation of \code{\link{auc}}, interpreted as the 
average difference between the distributions in standard deviation units.
}
\examples{
free_reduced <- rnorm(800, 80, 20)
pay <- rnorm(500, 100, 10)
d <- data.frame(score = c(free_reduced, pay), 
				frl = c(rep("free_reduced", 800),  
						rep("pay", 500)))

v(score ~ frl, d)
free_reduced <- rnorm(800, 80, 20)
pay <- rnorm(500, 100, 10)
d <- data.frame(score = c(free_reduced, pay), 
				frl = c(rep("free_reduced", 800),  
						rep("pay", 500)))

v(score ~ frl, d)
}

